+++ base/oix/interp.r
@@ -1549,7 +1549,7 @@ EntInterp;
 	    char buf[50];
 
 	    sprintf(buf, "unimplemented opcode: %ld (0x%08x)\n",
-               (long)lastop, lastop);
+                    (long)lastop, (long)lastop);
 	    syserr(buf);
 	    }
 	 }
